// src/prisma/generated/zod/inputTypeSchemas/TicketConfigurationMinOrderByAggregateInputSchema.ts
import { z } from 'zod';
import { type Prisma } from '@prisma/client';
import { SortOrderSchema } from './SortOrderSchema';

export const TicketConfigurationMinOrderByAggregateInputSchema: z.ZodType<Prisma.TicketConfigurationMinOrderByAggregateInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  jamId: z.lazy(() => SortOrderSchema).optional(),
  name: z.lazy(() => SortOrderSchema).optional(),
  description: z.lazy(() => SortOrderSchema).optional(),
  price: z.lazy(() => SortOrderSchema).optional(),
}).strict();

export default TicketConfigurationMinOrderByAggregateInputSchema;